Extract Expression Substitution code
Extracted substitution code from UI specific code so that
it can be used in other parts of the application. We would like
to use expressions with substitutions as Automate Method.
